// my-todo-app/src/generated/models/MicrosoftTo_Do_Business_Model.ts
export interface TodoList_V2 {
  "@odata.etag"?: string;
  // Unique identifier of the to-do list.
  id?: string;
  // Name of the to-do list.
  displayName?: string;
  // Property indicating the well-known list name if the given list is a well-known list. Possible values are: none, defaultList, flaggedEmails, unknownFutureValue.
  wellknownListName?: string;
  // True if the user is owner of the given task list.
  isOwner?: boolean;
  // True if the task list is shared with other users.
  isShared?: boolean;
}

export interface CreateToDoListBody {
  // List name
  name: string;
}

export type ToDo_V2importance = 'low'|'normal'|'high';
export type ToDo_V2status = 'notStarted'|'inProgress'|'completed'|'waitingOnOthers'|'deferred';

export interface ToDo_V2 {
  "@odata.context"?: string;
  "@odata.etag"?: string;
  // Unique identifier of the to-do.
  id?: string;
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  createdDateTime?: string;
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  lastModifiedDateTime?: string;
  // To-do body that typically contains information about the to-do.
  body?: {
    // The type of the content. Possible values are text and html.
    contentType?: string;
    // The content of the item.
    content?: string;
  };
  // YYYY-MM-DDThh:mm:ssZ (UTC format)
  bodyLastModifiedDateTime?: string;
  // The date in the specified time zone that the task was finished.
  completedDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // The date in the specified time zone that the task is to be finished.
  dueDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // Low, normal or high.
  importance?: ToDo_V2importance;
  // True if an alert is set to remind the user of the to-do.
  isReminderOn?: boolean;
  // The date and time for a reminder alert of the task to occur.
  reminderDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
  };
  // Indicates state or progress of the to-do - not started, in progress, completed, waiting on others or deferred.
  status?: ToDo_V2status;
  // Brief description or title of the to-do.
  title?: string;
}

export type CreateToDo_V2importance = 'low'|'normal'|'high';
export type CreateToDo_V2status = 'notStarted'|'inProgress'|'completed'|'waitingOnOthers'|'deferred';

export interface CreateToDo_V2 {
  // Date in the UTC time zone when the to-do is to be finished (note the time portion will be ignored).
  dueDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
    // Time zone of the due date.
    timeZone: string;
  };
  // Date-time in UTC time zone for a reminder alert of the to-do to occur.
  reminderDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
    // Time zone of the reminder date.
    timeZone: string;
  };
  // Low, normal or high.
  importance?: CreateToDo_V2importance;
  // Brief description or title of the to-do.
  title: string;
  // Indicates state or progress of the to-do - not started, in progress, completed, waiting on others or deferred.
  status?: CreateToDo_V2status;
  // To-do body that typically contains information about the to-do.
  body?: {
    // The type of the content. Possible values are text and html.
    contentType: string;
    // The content of the item.
    content?: string;
  };
  // True if an alert is set to remind the user of the to-do.
  isReminderOn?: boolean;
}

export type UpdateToDo_V2importance = 'low'|'normal'|'high';
export type UpdateToDo_V2status = 'notStarted'|'inProgress'|'completed'|'waitingOnOthers'|'deferred';

export interface UpdateToDo_V2 {
  // Date in the UTC time zone when the to-do is to be finished (note the time portion will be ignored).
  dueDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
    // Time zone of the due date.
    timeZone: string;
  };
  // Date-time in UTC time zone for a reminder alert of the to-do to occur.
  reminderDateTime?: {
    // YYYY-MM-DDThh:mm:ss
    dateTime?: string;
    // Time zone of the reminder date.
    timeZone: string;
  };
  // Low, normal or high.
  importance?: UpdateToDo_V2importance;
  // Brief description or title of the to-do.
  title?: string;
  // Indicates state or progress of the to-do - not started, in progress, completed, waiting on others or deferred.
  status?: UpdateToDo_V2status;
  // To-do body that typically contains information about the to-do.
  body?: {
    // The type of the content. Possible values are text and html.
    contentType: string;
    // The content of the item.
    content?: string;
  };
  // True if an alert is set to remind the user of the to-do.
  isReminderOn?: boolean;
}